data/requirements.txt
pandas
numpy
yfinance
ta         
fredapi
python-dotenv
scikit-learn
matplotlib
seaborn
scipy
transformers
torch
snscrape
joblib